+++ src/index.html
@@ -15,29 +15,7 @@
   </header>
 
   <section class="modal">
-    <!-- <div class="meal_photo">
-      <img src="https://static01.nyt.com/images/2020/03/25/dining/18clark-roast-chicken/clark-roast-chicken-superJumbo-v4.jpg" alt="roast chicken" height="300">
-      <i class="fas fa-times close_btn"></i>
-    </div>    
-    <h3 class="dish_title">Roast Chicken</h3>
-    <div class="dish_info">
-      <p>Category: Chicken</p>
-      <p>Main ingredient: Chicken</p>
-      <p>Area: US</p>
-      <p>Youtube:Link</p>
-    </div>
-    <div class="comments">
-      <h3 id="comment_title">Comments(2)</h3>
-      <p class="comment">Comment: Comment 1</p>
-      <p class="comment">Comment: Comment 2</p>
-      
-    </div>
-    <form action="#" id="comment_form">
-      <h3>Add a comment</h3>
-      <input type="text" placeholder="Your name" class="form_item" required>
-      <textarea name="insights" id="insight" cols="30" rows="10" class="form_item" placeholder="Your Comment" required></textarea>
-      <button type="submit" class="form_item" id="submit_btn">Comment</button>
-    </form> -->
+    
   </section>
 
   <footer id="footer">
